+++ qoqo/src/operations/mod.rs
@@ -188,6 +188,10 @@ pub fn operations(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     m.add_class::<InvSqrtPauliYWrapper>()?;
 
     // 1.16
+    m.add_class::<InvSGateWrapper>()?;
+    m.add_class::<InvTGateWrapper>()?;
+    m.add_class::<SXGateWrapper>()?;
+    m.add_class::<InvSXGateWrapper>()?;
     m.add_class::<TripleControlledPauliXWrapper>()?;
     m.add_class::<TripleControlledPauliZWrapper>()?;
     m.add_class::<TripleControlledPhaseShiftWrapper>()?;