+++ qoqo/tests/integration/operations/single_qubit_gate_operations.rs
@@ -17,10 +17,11 @@ use pyo3::prelude::*;
 use pyo3::Python;
 use qoqo::operations::convert_operation_to_pyobject;
 use qoqo::operations::{
-    GPi2Wrapper, GPiWrapper, HadamardWrapper, IdentityWrapper, InvSqrtPauliXWrapper,
-    InvSqrtPauliYWrapper, PauliXWrapper, PauliYWrapper, PauliZWrapper, PhaseShiftState0Wrapper,
-    PhaseShiftState1Wrapper, RotateAroundSphericalAxisWrapper, RotateXWrapper, RotateXYWrapper,
-    RotateYWrapper, RotateZWrapper, SGateWrapper, SingleQubitGateWrapper, SqrtPauliXWrapper,
+    GPi2Wrapper, GPiWrapper, HadamardWrapper, IdentityWrapper, InvSGateWrapper, InvSXGateWrapper,
+    InvSqrtPauliXWrapper, InvSqrtPauliYWrapper, InvTGateWrapper, PauliXWrapper, PauliYWrapper,
+    PauliZWrapper, PhaseShiftState0Wrapper, PhaseShiftState1Wrapper,
+    RotateAroundSphericalAxisWrapper, RotateXWrapper, RotateXYWrapper, RotateYWrapper,
+    RotateZWrapper, SGateWrapper, SXGateWrapper, SingleQubitGateWrapper, SqrtPauliXWrapper,
     SqrtPauliYWrapper, TGateWrapper,
 };
 use qoqo_calculator::Calculator;
@@ -219,6 +220,7 @@ fn test_new_invsqrtpaulix(input_operation: Operation, arguments: (u32,), method:
     })
 }
 
+/// Test new() function for SGate
 #[test_case(Operation::from(SGate::new(1)), (1,), "__eq__"; "SGate_eq")]
 #[test_case(Operation::from(SGate::new(1)), (0,), "__ne__"; "SGate_ne")]
 fn test_new_sgate(input_operation: Operation, arguments: (u32,), method: &str) {
@@ -254,6 +256,43 @@ fn test_new_sgate(input_operation: Operation, arguments: (u32,), method: &str) {
     })
 }
 
+/// Test new() function for InvSGate
+#[test_case(Operation::from(InvSGate::new(1)), (1,), "__eq__"; "InvSGate_eq")]
+#[test_case(Operation::from(InvSGate::new(1)), (0,), "__ne__"; "InvSGate_ne")]
+fn test_new_invsgate(input_operation: Operation, arguments: (u32,), method: &str) {
+    let operation = convert_operation_to_pyobject(input_operation).unwrap();
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation_type = py.get_type_bound::<InvSGateWrapper>();
+        let binding = operation_type.call1(arguments).unwrap();
+        let operation_py = binding.downcast::<InvSGateWrapper>().unwrap();
+
+        let comparison = bool::extract_bound(
+            &operation
+                .bind(py)
+                .call_method1(method, (operation_py,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison);
+
+        let def_wrapper = operation_py.extract::<InvSGateWrapper>().unwrap();
+        let binding = operation_type.call1((2,)).unwrap();
+        let new_op_diff = binding.downcast::<InvSGateWrapper>().unwrap();
+        let def_wrapper_diff = new_op_diff.extract::<InvSGateWrapper>().unwrap();
+        let helper_ne: bool = def_wrapper_diff != def_wrapper;
+        assert!(helper_ne);
+        let helper_eq: bool = def_wrapper == def_wrapper.clone();
+        assert!(helper_eq);
+
+        assert_eq!(
+            format!("{:?}", def_wrapper_diff),
+            "InvSGateWrapper { internal: InvSGate { qubit: 2 } }"
+        );
+    })
+}
+
+/// Test new() function for TGate
 #[test_case(Operation::from(TGate::new(1)), (1,), "__eq__"; "TGate_eq")]
 #[test_case(Operation::from(TGate::new(1)), (0,), "__ne__"; "TGate_ne")]
 fn test_new_tgate(input_operation: Operation, arguments: (u32,), method: &str) {
@@ -289,6 +328,43 @@ fn test_new_tgate(input_operation: Operation, arguments: (u32,), method: &str) {
     })
 }
 
+/// Test new() function for InvTGate
+#[test_case(Operation::from(InvTGate::new(1)), (1,), "__eq__"; "InvTGate_eq")]
+#[test_case(Operation::from(InvTGate::new(1)), (0,), "__ne__"; "InvTGate_ne")]
+fn test_new_invtgate(input_operation: Operation, arguments: (u32,), method: &str) {
+    let operation = convert_operation_to_pyobject(input_operation).unwrap();
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation_type = py.get_type_bound::<InvTGateWrapper>();
+        let binding = operation_type.call1(arguments).unwrap();
+        let operation_py = binding.downcast::<InvTGateWrapper>().unwrap();
+
+        let comparison = bool::extract_bound(
+            &operation
+                .bind(py)
+                .call_method1(method, (operation_py,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison);
+
+        let def_wrapper = operation_py.extract::<InvTGateWrapper>().unwrap();
+        let binding = operation_type.call1((2,)).unwrap();
+        let new_op_diff = binding.downcast::<InvTGateWrapper>().unwrap();
+        let def_wrapper_diff = new_op_diff.extract::<InvTGateWrapper>().unwrap();
+        let helper_ne: bool = def_wrapper_diff != def_wrapper;
+        assert!(helper_ne);
+        let helper_eq: bool = def_wrapper == def_wrapper.clone();
+        assert!(helper_eq);
+
+        assert_eq!(
+            format!("{:?}", def_wrapper_diff),
+            "InvTGateWrapper { internal: InvTGate { qubit: 2 } }"
+        );
+    })
+}
+
+/// Test new() function for Hadamard
 #[test_case(Operation::from(Hadamard::new(1)), (1,), "__eq__"; "Hadamard_eq")]
 #[test_case(Operation::from(Hadamard::new(1)), (0,), "__ne__"; "Hadamard_ne")]
 fn test_new_hadamard(input_operation: Operation, arguments: (u32,), method: &str) {
@@ -324,6 +400,7 @@ fn test_new_hadamard(input_operation: Operation, arguments: (u32,), method: &str
     })
 }
 
+/// Test new() function for RotateX
 #[test_case(Operation::from(RotateX::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "RotateX_eq")]
 #[test_case(Operation::from(RotateX::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "RotateX_ne")]
 fn test_new_rotatex(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -364,6 +441,7 @@ fn test_new_rotatex(input_operation: Operation, arguments: (u32, f64), method: &
     })
 }
 
+/// Test new() function for RotateY
 #[test_case(Operation::from(RotateY::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "RotateY_eq")]
 #[test_case(Operation::from(RotateY::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "RotateY_ne")]
 fn test_new_rotatey(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -404,6 +482,7 @@ fn test_new_rotatey(input_operation: Operation, arguments: (u32, f64), method: &
     })
 }
 
+/// Test new() function for RotateZ
 #[test_case(Operation::from(RotateZ::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "RotateZ_eq")]
 #[test_case(Operation::from(RotateZ::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "RotateZ_ne")]
 fn test_new_rotatez(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -445,6 +524,7 @@ fn test_new_rotatez(input_operation: Operation, arguments: (u32, f64), method: &
     })
 }
 
+/// Test new() function for PhaseShiftState0
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "PhaseShiftState0_eq")]
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "PhaseShiftState0_ne")]
 fn test_new_phaseshiftstate0(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -486,6 +566,7 @@ fn test_new_phaseshiftstate0(input_operation: Operation, arguments: (u32, f64), 
     })
 }
 
+/// Test new() function for PhaseShiftState1
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "PhaseShiftState1_eq")]
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "PhaseShiftState1_ne")]
 fn test_new_phaseshiftstate1(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -527,6 +608,7 @@ fn test_new_phaseshiftstate1(input_operation: Operation, arguments: (u32, f64), 
     })
 }
 
+/// Test new() function for GPi
 #[test_case(Operation::from(GPi::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "GPi_eq")]
 #[test_case(Operation::from(GPi::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "GPi_ne")]
 fn test_new_gpi(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -568,6 +650,7 @@ fn test_new_gpi(input_operation: Operation, arguments: (u32, f64), method: &str)
     })
 }
 
+/// Test new() function for GPi2
 #[test_case(Operation::from(GPi2::new(1, CalculatorFloat::ZERO)), (1, 0.0,), "__eq__"; "GPi2_eq")]
 #[test_case(Operation::from(GPi2::new(1, CalculatorFloat::ZERO)), (0, 0.0,), "__ne__"; "GPi2_ne")]
 fn test_new_gpi2(input_operation: Operation, arguments: (u32, f64), method: &str) {
@@ -913,6 +996,78 @@ fn test_new_invsqrtpauliy(input_operation: Operation, arguments: (u32,), method:
     })
 }
 
+/// Test new() function for SXGate
+#[test_case(Operation::from(SXGate::new(1)), (1,), "__eq__"; "SXGate_eq")]
+#[test_case(Operation::from(SXGate::new(1)), (0,), "__ne__"; "SXGate_ne")]
+fn test_new_sxgate(input_operation: Operation, arguments: (u32,), method: &str) {
+    let operation = convert_operation_to_pyobject(input_operation).unwrap();
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation_type = py.get_type_bound::<SXGateWrapper>();
+        let binding = operation_type.call1(arguments).unwrap();
+        let operation_py = binding.downcast::<SXGateWrapper>().unwrap();
+
+        let comparison = bool::extract_bound(
+            &operation
+                .bind(py)
+                .call_method1(method, (operation_py,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison);
+
+        let def_wrapper = operation_py.extract::<SXGateWrapper>().unwrap();
+        let binding = operation_type.call1((2,)).unwrap();
+        let new_op_diff = binding.downcast::<SXGateWrapper>().unwrap();
+        let def_wrapper_diff = new_op_diff.extract::<SXGateWrapper>().unwrap();
+        let helper_ne: bool = def_wrapper_diff != def_wrapper;
+        assert!(helper_ne);
+        let helper_eq: bool = def_wrapper == def_wrapper.clone();
+        assert!(helper_eq);
+
+        assert_eq!(
+            format!("{:?}", def_wrapper_diff),
+            "SXGateWrapper { internal: SXGate { qubit: 2 } }"
+        );
+    })
+}
+
+/// Test new() function for InvSXGate
+#[test_case(Operation::from(InvSXGate::new(1)), (1,), "__eq__"; "InvSXGate_eq")]
+#[test_case(Operation::from(InvSXGate::new(1)), (0,), "__ne__"; "InvSXGate_ne")]
+fn test_new_invsxgate(input_operation: Operation, arguments: (u32,), method: &str) {
+    let operation = convert_operation_to_pyobject(input_operation).unwrap();
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation_type = py.get_type_bound::<InvSXGateWrapper>();
+        let binding = operation_type.call1(arguments).unwrap();
+        let operation_py = binding.downcast::<InvSXGateWrapper>().unwrap();
+
+        let comparison = bool::extract_bound(
+            &operation
+                .bind(py)
+                .call_method1(method, (operation_py,))
+                .unwrap(),
+        )
+        .unwrap();
+        assert!(comparison);
+
+        let def_wrapper = operation_py.extract::<InvSXGateWrapper>().unwrap();
+        let binding = operation_type.call1((2,)).unwrap();
+        let new_op_diff = binding.downcast::<InvSXGateWrapper>().unwrap();
+        let def_wrapper_diff = new_op_diff.extract::<InvSXGateWrapper>().unwrap();
+        let helper_ne: bool = def_wrapper_diff != def_wrapper;
+        assert!(helper_ne);
+        let helper_eq: bool = def_wrapper == def_wrapper.clone();
+        assert!(helper_eq);
+
+        assert_eq!(
+            format!("{:?}", def_wrapper_diff),
+            "InvSXGateWrapper { internal: InvSXGate { qubit: 2 } }"
+        );
+    })
+}
+
 /// Test is_parametrized() function for SingleQubitGate Operations
 #[test_case(Operation::from(RotateX::new(0, CalculatorFloat::from("theta"))); "RotateX")]
 #[test_case(Operation::from(RotateY::new(0, CalculatorFloat::from("theta"))); "RotateY")]
@@ -1000,7 +1155,9 @@ fn test_pyo3_is_parametrized(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -1009,6 +1166,8 @@ fn test_pyo3_is_parametrized(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(1)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(3)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(3)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(100)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(100)); "InvSXGate")]
 fn test_pyo3_is_not_parametrized(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1083,7 +1242,9 @@ fn test_pyo3_theta(theta: CalculatorFloat, input_operation: Operation) {
 #[test_case(0, Operation::from(SqrtPauliX::new(0)); "SqrtPauliX")]
 #[test_case(0, Operation::from(InvSqrtPauliX::new(0)); "InvSqrtPauliX")]
 #[test_case(0, Operation::from(TGate::new(0)); "TGate")]
+#[test_case(0, Operation::from(InvTGate::new(0)); "InvTGate")]
 #[test_case(0, Operation::from(SGate::new(0)); "SGate")]
+#[test_case(0, Operation::from(InvSGate::new(0)); "InvSGate")]
 #[test_case(0, Operation::from(Hadamard::new(0)); "Hadamard")]
 #[test_case(1, Operation::from(PhaseShiftState0::new(1, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(1, Operation::from(PhaseShiftState1::new(1, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -1092,6 +1253,8 @@ fn test_pyo3_theta(theta: CalculatorFloat, input_operation: Operation) {
 #[test_case(1, Operation::from(Identity::new(1)); "Identity")]
 #[test_case(3, Operation::from(SqrtPauliY::new(3)); "SqrtPauliY")]
 #[test_case(3, Operation::from(InvSqrtPauliY::new(3)); "InvSqrtPauliY")]
+#[test_case(0, Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(0, Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_qubit(qubit: usize, input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1134,7 +1297,9 @@ fn test_pyo3_qubit(qubit: usize, input_operation: Operation) {
 #[test_case("SqrtPauliX", Operation::from(SqrtPauliX::new(0)); "SqrtPauliX")]
 #[test_case("InvSqrtPauliX", Operation::from(InvSqrtPauliX::new(0)); "InvSqrtPauliX")]
 #[test_case("SGate", Operation::from(SGate::new(0)); "SGate")]
+#[test_case("InvSGate", Operation::from(InvSGate::new(0)); "InvSGate")]
 #[test_case("TGate", Operation::from(TGate::new(0)); "TGate")]
+#[test_case("InvTGate", Operation::from(InvTGate::new(0)); "InvTGate")]
 #[test_case("Hadamard", Operation::from(Hadamard::new(0)); "Hadamard")]
 #[test_case("SingleQubitGate", Operation::from(
     SingleQubitGate::new(
@@ -1152,6 +1317,8 @@ fn test_pyo3_qubit(qubit: usize, input_operation: Operation) {
 #[test_case("Identity", Operation::from(Identity::new(1)); "Identity")]
 #[test_case("SqrtPauliY", Operation::from(SqrtPauliY::new(3)); "SqrtPauliY")]
 #[test_case("InvSqrtPauliY", Operation::from(InvSqrtPauliY::new(3)); "InvSqrtPauliY")]
+#[test_case("SXGate", Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case("InvSXGate", Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_hqslang(name: &'static str, input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1224,6 +1391,15 @@ fn test_pyo3_hqslang(name: &'static str, input_operation: Operation) {
         ];
     "TGate")]
 #[test_case(
+    Operation::from(InvTGate::new(0)),
+    vec![
+        "Operation",
+        "GateOperation",
+        "SingleQubitGateOperation",
+        "InvTGate",
+        ];
+    "InvTGate")]
+#[test_case(
     Operation::from(SGate::new(0)),
     vec![
         "Operation",
@@ -1232,6 +1408,15 @@ fn test_pyo3_hqslang(name: &'static str, input_operation: Operation) {
         "SGate",
         ];
     "SGate")]
+#[test_case(
+    Operation::from(InvSGate::new(0)),
+    vec![
+        "Operation",
+        "GateOperation",
+        "SingleQubitGateOperation",
+        "InvSGate",
+        ];
+    "InvSGate")]
 #[test_case(
     Operation::from(PauliX::new(0)),
     vec![
@@ -1369,6 +1554,24 @@ fn test_pyo3_hqslang(name: &'static str, input_operation: Operation) {
         "InvSqrtPauliY",
         ];
     "InvSqrtPauliY")]
+#[test_case(
+    Operation::from(SXGate::new(0)),
+    vec![
+        "Operation",
+        "GateOperation",
+        "SingleQubitGateOperation",
+        "SXGate",
+        ];
+    "SXGate")]
+#[test_case(
+    Operation::from(InvSXGate::new(0)),
+    vec![
+        "Operation",
+        "GateOperation",
+        "SingleQubitGateOperation",
+        "InvSXGate",
+        ];
+    "InvSXGate")]
 fn test_pyo3_tags(input_operation: Operation, tags: Vec<&str>) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1424,7 +1627,9 @@ fn test_pyo3_tags(input_operation: Operation, tags: Vec<&str>) {
 #[test_case(Operation::from(SqrtPauliX::new(0)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(0)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(0)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(0)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(0)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(0)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(0)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -1433,6 +1638,8 @@ fn test_pyo3_tags(input_operation: Operation, tags: Vec<&str>) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_remapqubits(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1503,7 +1710,9 @@ fn test_pyo3_remapqubits(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(0)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(0)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(0)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(0)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(0)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(0)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(0)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -1512,6 +1721,8 @@ fn test_pyo3_remapqubits(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_remapqubits_error(input_operation: Operation) {
     // preparation
     pyo3::prepare_freethreaded_python();
@@ -1552,7 +1763,9 @@ fn test_pyo3_remapqubits_error(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(
     SingleQubitGate::new(
@@ -1572,6 +1785,8 @@ fn test_pyo3_remapqubits_error(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_unitarymatrix(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1698,7 +1913,9 @@ fn test_pyo3_unitarymatrix_singlequbitgate(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -1707,6 +1924,8 @@ fn test_pyo3_unitarymatrix_singlequbitgate(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_copy_deepcopy(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1772,7 +1991,9 @@ fn test_pyo3_copy_deepcopy(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -1781,6 +2002,8 @@ fn test_pyo3_copy_deepcopy(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_alpha_r(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1840,7 +2063,9 @@ fn test_pyo3_alpha_r(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -1849,6 +2074,8 @@ fn test_pyo3_alpha_r(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_alpha_i(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1908,7 +2135,9 @@ fn test_pyo3_alpha_i(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -1917,6 +2146,8 @@ fn test_pyo3_alpha_i(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_beta_r(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -1976,7 +2207,9 @@ fn test_pyo3_beta_r(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -1985,6 +2218,8 @@ fn test_pyo3_beta_r(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_beta_i(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2044,7 +2279,9 @@ fn test_pyo3_beta_i(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -2053,6 +2290,8 @@ fn test_pyo3_beta_i(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_global_phase(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2113,9 +2352,17 @@ fn test_pyo3_global_phase(input_operation: Operation) {
     Operation::from(SGate::new(0));
     "SGate")]
 #[test_case(
+    "InvSGate { qubit: 0 }",
+    Operation::from(InvSGate::new(0));
+    "InvSGate")]
+#[test_case(
     "TGate { qubit: 0 }",
     Operation::from(TGate::new(0));
     "TGate")]
+#[test_case(
+    "InvTGate { qubit: 0 }",
+    Operation::from(InvTGate::new(0));
+    "InvTGate")]
 #[test_case(
     "Hadamard { qubit: 0 }",
     Operation::from(Hadamard::new(0));
@@ -2171,6 +2418,14 @@ fn test_pyo3_global_phase(input_operation: Operation) {
     "InvSqrtPauliY { qubit: 0 }",
     Operation::from(InvSqrtPauliY::new(0));
     "InvSqrtPauliY")]
+#[test_case(
+    "SXGate { qubit: 0 }",
+    Operation::from(SXGate::new(0));
+    "SXGate")]
+#[test_case(
+    "InvSXGate { qubit: 0 }",
+    Operation::from(InvSXGate::new(0));
+    "InvSXGate")]
 fn test_pyo3_format_repr(format_repr: &str, input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2316,7 +2571,9 @@ fn test_pyo3_substitute_params_error(input_operation: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -2328,6 +2585,8 @@ fn test_pyo3_substitute_params_error(input_operation: Operation) {
 #[test_case(Operation::from(Identity::new(0)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_ineffective_substitute_parameters(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2421,7 +2680,9 @@ fn test_pyo3_rotate_powercf(first_op: Operation, second_op: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(1)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(1)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(1)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -2433,6 +2694,8 @@ fn test_pyo3_rotate_powercf(first_op: Operation, second_op: Operation) {
 #[test_case(Operation::from(Identity::new(1)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(1)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(1)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(1)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(1)); "InvSXGate")]
 fn test_pyo3_mul(gate1: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2459,7 +2722,9 @@ fn test_pyo3_mul(gate1: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(1)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(1)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(1)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -2471,6 +2736,8 @@ fn test_pyo3_mul(gate1: Operation) {
 #[test_case(Operation::from(Identity::new(1)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(1)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(1)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(1)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(1)); "InvSXGate")]
 fn test_pyo3_mul_error1(gate1: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2490,7 +2757,9 @@ fn test_pyo3_mul_error1(gate1: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(1)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(1)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(1)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -2502,6 +2771,8 @@ fn test_pyo3_mul_error1(gate1: Operation) {
 #[test_case(Operation::from(Identity::new(1)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(0)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_mul_error2(gate1: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2521,7 +2792,9 @@ fn test_pyo3_mul_error2(gate1: Operation) {
 #[test_case(Operation::from(SqrtPauliX::new(1)); "SqrtPauliX")]
 #[test_case(Operation::from(InvSqrtPauliX::new(1)); "InvSqrtPauliX")]
 #[test_case(Operation::from(SGate::new(1)); "SGate")]
+#[test_case(Operation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(Operation::from(TGate::new(1)); "TGate")]
+#[test_case(Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(Operation::from(Hadamard::new(1)); "Hadamard")]
 #[test_case(Operation::from(PhaseShiftState0::new(1, CalculatorFloat::from(0))); "PhaseShiftState0")]
 #[test_case(Operation::from(PhaseShiftState1::new(1, CalculatorFloat::from(0))); "PhaseShiftState1")]
@@ -2533,6 +2806,8 @@ fn test_pyo3_mul_error2(gate1: Operation) {
 #[test_case(Operation::from(Identity::new(1)); "Identity")]
 #[test_case(Operation::from(SqrtPauliY::new(1)); "SqrtPauliY")]
 #[test_case(Operation::from(InvSqrtPauliY::new(1)); "InvSqrtPauliY")]
+#[test_case(Operation::from(SXGate::new(1)); "SXGate")]
+#[test_case(Operation::from(InvSXGate::new(1)); "InvSXGate")]
 fn test_pyo3_mul_error3(gate1: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2599,8 +2874,14 @@ fn test_pyo3_mul_error3(gate1: Operation) {
     Operation::from(SGate::new(0)),
     Operation::from(SGate::new(1)); "SGate")]
 #[test_case(
+    Operation::from(InvSGate::new(0)),
+    Operation::from(InvSGate::new(1)); "InvSGate")]
+#[test_case(
     Operation::from(TGate::new(0)),
     Operation::from(TGate::new(1)); "TGate")]
+#[test_case(
+    Operation::from(InvTGate::new(0)),
+    Operation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(
     Operation::from(Hadamard::new(0)),
     Operation::from(Hadamard::new(1)); "Hadamard")]
@@ -2647,6 +2928,12 @@ fn test_pyo3_mul_error3(gate1: Operation) {
 #[test_case(
     Operation::from(InvSqrtPauliY::new(0)),
     Operation::from(InvSqrtPauliY::new(1)); "InvSqrtPauliY")]
+#[test_case(
+    Operation::from(SXGate::new(0)),
+    Operation::from(SXGate::new(1)); "SXGate")]
+#[test_case(
+    Operation::from(InvSXGate::new(0)),
+    Operation::from(InvSXGate::new(1)); "InvSXGate")]
 fn test_pyo3_richcmp(definition_1: Operation, definition_2: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -2718,7 +3005,9 @@ fn test_pyo3_richcmp(definition_1: Operation, definition_2: Operation) {
 #[test_case(SingleQubitGateOperation::from(SqrtPauliX::new(100)); "SqrtPauliX")]
 #[test_case(SingleQubitGateOperation::from(InvSqrtPauliX::new(100)); "InvSqrtPauliX")]
 #[test_case(SingleQubitGateOperation::from(SGate::new(1)); "SGate")]
+#[test_case(SingleQubitGateOperation::from(InvSGate::new(1)); "InvSGate")]
 #[test_case(SingleQubitGateOperation::from(TGate::new(1)); "TGate")]
+#[test_case(SingleQubitGateOperation::from(InvTGate::new(1)); "InvTGate")]
 #[test_case(SingleQubitGateOperation::from(Hadamard::new(3)); "Hadamard")]
 #[test_case(SingleQubitGateOperation::from(PhaseShiftState0::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState0")]
 #[test_case(SingleQubitGateOperation::from(PhaseShiftState1::new(0, CalculatorFloat::from(0.0))); "PhaseShiftState1")]
@@ -2727,6 +3016,8 @@ fn test_pyo3_richcmp(definition_1: Operation, definition_2: Operation) {
 #[test_case(SingleQubitGateOperation::from(Identity::new(1)); "Identity")]
 #[test_case(SingleQubitGateOperation::from(SqrtPauliY::new(0)); "SqrtPauliY")]
 #[test_case(SingleQubitGateOperation::from(InvSqrtPauliY::new(0)); "InvSqrtPauliY")]
+#[test_case(SingleQubitGateOperation::from(SXGate::new(0)); "SXGate")]
+#[test_case(SingleQubitGateOperation::from(InvSXGate::new(0)); "InvSXGate")]
 fn test_pyo3_json_schema(operation: SingleQubitGateOperation) {
     let rust_schema = match operation {
         SingleQubitGateOperation::SingleQubitGate(_) => {
@@ -2762,8 +3053,14 @@ fn test_pyo3_json_schema(operation: SingleQubitGateOperation) {
         SingleQubitGateOperation::SGate(_) => {
             serde_json::to_string_pretty(&schemars::schema_for!(SGate)).unwrap()
         }
+        SingleQubitGateOperation::InvSGate(_) => {
+            serde_json::to_string_pretty(&schemars::schema_for!(InvSGate)).unwrap()
+        }
         SingleQubitGateOperation::TGate(_) => {
             serde_json::to_string_pretty(&schemars::schema_for!(TGate)).unwrap()
+        }
+        SingleQubitGateOperation::InvTGate(_) => {
+            serde_json::to_string_pretty(&schemars::schema_for!(InvTGate)).unwrap()
         }
         SingleQubitGateOperation::PhaseShiftState1(_) => {
             serde_json::to_string_pretty(&schemars::schema_for!(PhaseShiftState1)).unwrap()
@@ -2792,6 +3089,12 @@ fn test_pyo3_json_schema(operation: SingleQubitGateOperation) {
         SingleQubitGateOperation::InvSqrtPauliY(_) => {
             serde_json::to_string_pretty(&schemars::schema_for!(InvSqrtPauliY)).unwrap()
         }
+        SingleQubitGateOperation::SXGate(_) => {
+            serde_json::to_string_pretty(&schemars::schema_for!(SXGate)).unwrap()
+        }
+        SingleQubitGateOperation::InvSXGate(_) => {
+            serde_json::to_string_pretty(&schemars::schema_for!(InvSXGate)).unwrap()
+        }
         _ => unreachable!(),
     };
     pyo3::prepare_freethreaded_python();
@@ -2802,6 +3105,10 @@ fn test_pyo3_json_schema(operation: SingleQubitGateOperation) {
             SingleQubitGateOperation::Identity(_) => "1.7.0".to_string(),
             SingleQubitGateOperation::SqrtPauliY(_) => "1.15.0".to_string(),
             SingleQubitGateOperation::InvSqrtPauliY(_) => "1.15.0".to_string(),
+            SingleQubitGateOperation::InvSGate(_)
+            | SingleQubitGateOperation::InvTGate(_)
+            | SingleQubitGateOperation::SXGate(_)
+            | SingleQubitGateOperation::InvSXGate(_) => "1.16.0".to_string(),
             _ => "1.0.0".to_string(),
         };
         let converted_op = Operation::from(operation);